Add Firebase multi-device foundation with guest sync
Introduce a provider-agnostic data layer so the app can run either
fully local (Hive) or cloud-backed (Firestore) without the UI knowing
the difference, plus the money/settlement groundwork it builds on.

Data layer
- ExpensioRepository interface with Hive and Firestore implementations
- AppState (ChangeNotifier) view-model caching repository streams and
  exposing synchronous getters; screens migrated off direct Hive access
- Pure computeBalancesFrom() shared by both backends

Firebase
- Anonymous guest auth with guarded init: falls back to local-only mode
  when Firebase is not configured
- Firestore offline persistence as the local cache
- Member-only security rules (groups by memberIds, expenses by visibleTo,
  settlements by denormalized memberIds)
- Signed-in user gets a "You" profile keyed by their auth uid and is
  always a member of groups they create; group expenses are visible to
  all group members

Money and settlements
- Store amounts as integer minor units (cents) to avoid float drift,
  with exact remainder distribution on equal splits
- Persist settle-up records; balances account for them
- Cap settlement backtracking to avoid blow-up on larger groups

Platform
- Android minSdk 23, iOS 13 (Firebase requirements)

+++ lib/models/expense_model.dart
@@ -11,8 +11,9 @@ class ExpenseModel extends HiveObject {
   @HiveField(1)
   String title;
 
+  /// Total amount in minor units (cents/paise).
   @HiveField(2)
-  double totalAmount;
+  int totalAmount;
 
   @HiveField(3)
   String payerId;
@@ -35,9 +36,9 @@ class ExpenseModel extends HiveObject {
   @HiveField(9)
   bool isPersonal;
 
-  // Split map: userId -> amount they owe
+  // Split map: userId -> amount they owe, in minor units (cents/paise)
   @HiveField(10)
-  Map<String, double> splitMap;
+  Map<String, int> splitMap;
 
   ExpenseModel({
     required this.id,
@@ -50,7 +51,7 @@ class ExpenseModel extends HiveObject {
     List<BillItem>? items,
     this.category = 'General',
     this.isPersonal = false,
-    Map<String, double>? splitMap,
+    Map<String, int>? splitMap,
   })  : items = items ?? [],
         splitMap = splitMap ?? {};
 }
